+++ Assets/Scripts/Combat/Test_HitResponder.cs
@@ -6,10 +6,10 @@ namespace Combat
     public class Test_HitResponder : MonoBehaviour, IHitResponder
     {
         [SerializeField] private bool m_attack;
-        [SerializeField] private int m_damage = 10;
+        [SerializeField] private float m_damage = 10;
         [SerializeField] private Comp_Hitbox _hitbox;
         
-        int IHitResponder.Damage { get => m_damage; }
+        float IHitResponder.Damage { get => m_damage; }
 
         private void Start()
         {
